% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate Entropy Impurity}
\usage{
calculate_entropy(y)
}
\arguments{
\item{y}{Binary target vector (0/1).}
}
\value{
Numeric entropy impurity value (in bits).
}
\description{
Computes the entropy impurity of a binary vector. Useful for decision tree splitting criteria.
}
\examples{
y = c(0, 1, 1, 0, 1)
calculate_entropy(y)

}
