% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{calculate_gini}
\alias{calculate_gini}
\title{Calculate Gini Impurity}
\usage{
calculate_gini(y)
}
\arguments{
\item{y}{Binary target vector (0/1).}
}
\value{
Numeric Gini impurity value.
}
\description{
Computes the Gini impurity of a binary vector. Useful for decision tree splitting criteria.
}
\examples{
y = c(0, 1, 1, 0, 1)
calculate_gini(y)

}
