% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{calculate_impurity}
\alias{calculate_impurity}
\title{Calculate Impurity Based on Specified Criterion}
\usage{
calculate_impurity(y, criterion = "gini")
}
\arguments{
\item{y}{Binary target vector (0/1).}

\item{criterion}{Impurity criterion to use: "gini" or "entropy".}
}
\value{
Numeric impurity value.
}
\description{
Computes impurity (Gini or Entropy) of a binary target vector based on the selected criterion.
}
\examples{
y = c(0, 1, 1, 0, 1)
calculate_impurity(y, "gini")
calculate_impurity(y, "entropy")

}
