% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{dataPreprocessing}
\alias{dataPreprocessing}
\title{Preprocess a Data Frame for Machine Learning}
\usage{
dataPreprocessing(data, one_hot = TRUE, split_date = TRUE, label_col = NULL)
}
\arguments{
\item{data}{A `data.frame`. The input dataset to be preprocessed.}

\item{one_hot}{Logical, default is `TRUE`. Whether to apply one-hot encoding to character and factor columns.}

\item{split_date}{Logical, default is `TRUE`. Whether to split a `Date` column into `Year`, `Month`, and `Day`.}

\item{label_col}{Character, default is `NULL`. The name of the label (target) column. This column will be moved to the last position.}
}
\value{
A preprocessed `data.frame` with missing rows removed, optional one-hot encoding applied,
        and the label column moved to the end. If `split_date = TRUE`, the `Date` column is removed
        and replaced with three columns: `Year`, `Month`, and `Day`.
}
\description{
This function cleans and preprocesses a dataset by removing missing values,
optionally splitting a `Date` column into `Year`, `Month`, and `Day`,
and performing one-hot encoding for categorical variables. It also allows
specifying a label column which will be moved to the last column of the returned data.
}
\details{
The function performs several preprocessing steps in the following order:
\enumerate{
  \item Asserts that `data` is a data frame.
  \item Removes all rows containing any `NA` values.
  \item If `split_date` is `TRUE`, attempts to parse the `Date` column and split it into `Year`, `Month`, and `Day`.
  \item If `one_hot` is `TRUE`, one-hot encodes all categorical (character or factor) columns, except for the label column.
  \item The label column, if specified, is moved to the end of the dataset.
}

If `label_col` is a categorical variable and `one_hot = TRUE`, then it must produce exactly one column during one-hot encoding.
}
\examples{
# Example with categorical variable and label column
df <- data.frame(A = c("Yes", "No", "Yes"), B = 1:3)
dataPreprocessing(df, one_hot = TRUE, split_date = FALSE, label_col = "A")

# Example with date splitting
df2 <- data.frame(Date = c("2020-01-01", "2021-03-15"), X = 1:2)
dataPreprocessing(df2, one_hot = FALSE, split_date = TRUE)

# Example with NA removal
df3 <- data.frame(A = c(1, NA), B = c("x", "y"))
\dontrun{
dataPreprocessing(df3)  # Throws an error due to all rows removed
}

}
\seealso{
\code{\link[assertthat]{assert_that}}, \code{\link[lubridate]{ymd}}
}
