% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{find_best_split_for_feature}
\alias{find_best_split_for_feature}
\title{Find the Best Split Point for a Single Feature}
\usage{
find_best_split_for_feature(x, y, criterion = "gini")
}
\arguments{
\item{x}{A numeric vector representing a single feature.}

\item{y}{A binary target vector (0/1).}

\item{criterion}{Impurity criterion to use, either "gini" or "entropy".}
}
\value{
A list containing:
\describe{
  \item{split}{The best split point (numeric).}
  \item{impurity}{The weighted impurity for this split.}
}
}
\description{
Identifies the optimal threshold to split a numeric feature by minimizing impurity (Gini or Entropy).
}
\examples{
x = c(1, 2, 3, 4, 5)
y = c(0, 0, 1, 1, 1)
find_best_split_for_feature(x, y, criterion = "gini")

}
