% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{predict_tree_batch}
\alias{predict_tree_batch}
\title{Predict Multiple Samples Using the Decision Tree}
\usage{
predict_tree_batch(tree, X)
}
\arguments{
\item{tree}{A trained decision tree.}

\item{X}{Feature data.frame.}
}
\value{
A vector of predicted classes.
}
\description{
Predicts binary class labels for a batch of samples using a trained decision tree.
Each row in the input data is treated as a separate observation.
}
\examples{
# Create toy data
X = data.frame(
  x1 = c(2, 5, 4, 3, 7, 2),
  x2 = c(1, 2, 1, 1, 3, 2)
)
y = c(0, 0, 1, 1, 1, 0)

# Train a decision tree
tree = build_tree(X, y, max_depth = 2, min_samples_split = 2, criterion = "gini")

# Predict labels for the data
preds = predict_tree_batch(tree, X)
print(preds)

}
