% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{predict_tree_prob}
\alias{predict_tree_prob}
\title{Predict Probability for a Single Observation Using Decision Tree}
\usage{
predict_tree_prob(tree, x)
}
\arguments{
\item{tree}{A trained decision tree.}

\item{x}{A named list or vector representing one sample.}
}
\value{
Predicted probability of y = 1
}
\description{
Traverses a trained decision tree to predict the probability of the positive class (y = 1) for a single observation.
Returns the probability value stored in the corresponding leaf node.
}
\examples{
# Create toy data
X = data.frame(
  x1 = c(2, 5, 4, 3, 7, 2),
  x2 = c(1, 2, 1, 1, 3, 2)
)
y = c(0, 0, 1, 1, 1, 0)

# Train a decision tree
tree = build_tree(X, y, max_depth = 2, min_samples_split = 2)

# Predict probability for a single sample
new_sample = as.list(X[1, ])
prob = predict_tree_prob(tree, new_sample)
print(prob)

}
