% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{train_random_forest}
\alias{train_random_forest}
\title{Train a Random Forest Model}
\usage{
train_random_forest(X, y, ntree = 100, mtry = NULL, nodesize = 1)
}
\arguments{
\item{X}{Feature data.frame.}

\item{y}{Target vector.}

\item{ntree}{Number of trees.}

\item{mtry}{Number of features to consider at each split.}

\item{nodesize}{Minimum size of terminal nodes.}
}
\value{
Trained random forest model.
}
\description{
Trains a random forest classifier on a binary outcome using the \code{randomForest} package.
Allows control of number of trees, number of features considered per split, and terminal node size.
}
\examples{
# Create toy data
X = data.frame(
  x1 = c(1, 2, 3, 4),
  x2 = c(5, 6, 7, 8)
)
y = c(0, 0, 1, 1)

# Train a random forest model
model = train_random_forest(X, y, ntree = 50)

}
